+++ wde/__init__.py
@@ -1,3 +1,3 @@
 """Web Design Enhancer V3 core package."""
 
-__version__ = "3.0.0a1"
+__version__ = "3.0.0a4"